+++ src/boututils/upgrader/bout_v5_header_upgrader.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-import boutupgrader
+from .common import default_args, apply_or_display_patch
 
 import argparse
 import copy
@@ -120,7 +120,7 @@ if __name__ == "__main__":
             """
         ),
     )
-    parser = boutupgrader.default_args(parser, pos_files=False)
+    parser = default_args(parser, pos_files=False)
 
     parser.add_argument(
         "--include-path",
@@ -158,6 +158,6 @@ if __name__ == "__main__":
 
         modified = apply_fixes(header_regex, contents)
 
-        boutupgrader.apply_or_display_patch(
+        apply_or_display_patch(
             filename, original, modified, args.patch_only, args.quiet, args.force
         )
